+++ packages/pluggableWidgets/switch-web/src/Switch.editorConfig.ts
@@ -2,6 +2,7 @@ import { StructurePreviewProps } from "@mendix/pluggable-widgets-commons";
 
 import checkedSVG from "./assets/checked.svg";
 import checkedSVGDark from "./assets/checked-dark.svg";
+import { SwitchPreviewProps } from "typings/SwitchProps";
 
 export function getPreview(_: StructurePreviewProps, isDarkMode: boolean): StructurePreviewProps {
     return {
@@ -9,4 +10,8 @@ export function getPreview(_: StructurePreviewProps, isDarkMode: boolean): Struc
         document: decodeURIComponent((isDarkMode ? checkedSVGDark : checkedSVG).replace("data:image/svg+xml,", "")),
         width: 80
     };
+}
+
+export function getCustomCaption(values: SwitchPreviewProps, _platform = "desktop"): string {
+    return values.booleanAttribute || `Switch`;
 }