+++ packages/pluggableWidgets/rating-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   We updated page explorer's caption to display datasource.
+
 ## [3.1.2] - 2023-05-23
 
 ### Changed